Add-on to fix "module" name not found
There is an issue where the original v1 QRC generator script cannot have "import [non-bpy-module]", thus the pip check section of the script need to be seperated out as independent addon, thus the library and modules can be imported and used.  - This addon is a ugly quick fix, should update soon.